feat : api 호출 및 geport id전달 추가 fix : 파일 구조 변경 및 라우터 변경
