+++ package.json
@@ -3,12 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "node server.js",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
+    "express": "^4.19.2",
+    "http-proxy-middleware": "^3.0.0",
     "next": "14.1.4",
     "react": "^18",
     "react-dom": "^18"
